+++ src/app/layout.tsx
@@ -1,6 +1,7 @@
 import type { Metadata } from 'next';
 import './globals.css';
 import Providers from '@/config/tq-provider';
+import Header from '@/components/layouts/header-temp';
 
 export const metadata: Metadata = {
   title: '제주 올레요',
@@ -14,6 +15,7 @@ export default function RootLayout({
   return (
     <html lang="en">
       <body>
+        <Header />
         <Providers>{children}</Providers>
       </body>
     </html>
